Extract out fps calculation logic
